tests/turbohtml-tests/xml_coercion.dat

#data
#xml-coercion
<!-- -- -->
#errors
#document
| <!--  - -  -->
| <html>
|   <head>
|   <body>

#data
#xml-coercion
<div></div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       " "
